Fix adjustDate to use long for microseconds

+++ legend-pure-m4/src/main/java/org/finos/legend/pure/m4/coreinstance/primitive/date/DateWithHour.java
@@ -79,7 +79,7 @@ public final class DateWithHour extends AbstractDateWithHour
     }
 
     @Override
-    public PureDate addMicroseconds(int microseconds)
+    public PureDate addMicroseconds(long microseconds)
     {
         throw new UnsupportedOperationException();
     }
