+++ legend-pure-runtime-java-engine-compiled/src/main/java/org/finos/legend/pure/runtime/java/compiled/generation/processors/support/CompiledSupport.java
@@ -2729,7 +2729,7 @@ public class CompiledSupport
             }
             case "MICROSECONDS":
             {
-                return date.addMicroseconds((int)number);
+                return date.addMicroseconds(number);
             }
             case "NANOSECONDS":
             {
